refactor(console): update per review
