--- packages/console/src/components/AppContent/components/Content/index.module.scss
@@ -1,3 +1,0 @@
-.content {
-  flex-grow: 1;
-}
